Make gossip scores available to users - code review changes
