+++ data/serializer/src/main/java/tech/pegasys/teku/api/response/v1/teku/GetPeerScoresResponse.java
@@ -1,0 +1,27 @@
+/*
+ * Copyright 2021 ConsenSys AG.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+package tech.pegasys.teku.api.response.v1.teku;
+
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import java.util.List;
+
+public class GetPeerScoresResponse {
+  public final List<PeerScore> data;
+
+  @JsonCreator
+  public GetPeerScoresResponse(@JsonProperty("data") final List<PeerScore> data) {
+    this.data = data;
+  }
+}
