+++ networking/p2p/src/main/java/tech/pegasys/teku/networking/p2p/libp2p/LibP2PNetwork.java
@@ -114,7 +114,9 @@ public class LibP2PNetwork implements P2PNetwork<Peer> {
     rpcMethods.forEach(method -> rpcHandlers.put(method, new RpcHandler(asyncRunner, method)));
 
     // Setup peers
-    peerManager = new PeerManager(metricsSystem, reputationManager, peerHandlers, rpcHandlers, gossipNetwork.getGossip());
+    peerManager =
+        new PeerManager(
+            metricsSystem, reputationManager, peerHandlers, rpcHandlers, gossipNetwork.getGossip());
 
     final Multiaddr listenAddr =
         MultiaddrUtil.fromInetSocketAddress(
